requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apispec==6.4.0
apprise==1.9.3
APScheduler==3.11.0
argcomplete==3.2.1
asgi-lifespan==2.1.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
Babel==2.14.0
backoff==2.2.1
blinker==1.7.0
cachelib==0.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
coolname==2.2.0
cron_descriptor==1.4.0
croniter==2.0.1
cryptography==44.0.2
cycler==0.12.1
dateparser==1.2.1
Deprecated==1.2.18
dill==0.3.1.1
docker==7.1.0
docutils==0.20.1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
flasgger==0.9.7.1
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Caching==2.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==25.1.24
fonttools==4.55.8
frozenlist==1.4.1
fsspec==2025.3.2
gast==0.6.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.2
hyperframe==6.1.0
idna==3.10
ImageHash==4.3.2
importlib-resources==6.1.1
importlib_metadata==8.6.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.8.0
keybert==0.9.0
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
libclang==18.1.1
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.2
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.0.4
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.1.0
orjson==3.10.16
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pgvector==0.4.1
pillow==11.1.0
pluggy==1.3.0
prefect==3.3.7
prison==0.2.1
prometheus_client==0.21.1
protobuf==4.25.2
psutil==5.9.7
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nvd3==0.15.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
readchar==4.2.1
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
text-unidecode==1.3
tf_keras==2.19.0
threadpoolctl==3.6.0
time-machine==2.13.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.2
ujson==5.10.0
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==2.4.0
uv==0.6.17
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==2.2.3
